uVault-main/requirements.txt
Pillow==8.1.0
opencv-python==4.5.5.64
azure-cognitiveservices-vision-face==0.6.0
azure-common==1.1.28